added LCM, GCD; refactored many and updated 'DOCUMENTATION' for all

+++ Java Projects/10 SimpleOrCompundGrowth/src/CompGrowthAnnual.java
@@ -1,9 +1,9 @@
 class CompGrowthAnnual {
-    float growthRate;
-    long principal;
-    float numOfYears;
+    double growthRate;
+    double principal;
+    double numOfYears;
 
-    public CompGrowthAnnual(float growthRate, long principal, float numOfYears) {
+    public CompGrowthAnnual(double growthRate, double principal, double numOfYears) {
         this.growthRate = growthRate;
         this.principal = principal;
         this.numOfYears = numOfYears;
